+++ apps/ced-card-request-fe/src/features/photo-upload/reducer.ts
@@ -3,6 +3,7 @@ import { RootState } from '../../core/store';
 
 type PhotoState = {
   B64_photo?: string;
+  preview?: string;
 };
 
 const initialState: PhotoState = {};
@@ -14,11 +15,15 @@ const photoSlice = createSlice({
     setFile: (state, action: PayloadAction<string>) => {
       state.B64_photo = action.payload;
     },
+    setPreview: (state, action: PayloadAction<string | undefined>) => {
+      state.preview = action.payload;
+    },
   },
 });
 
-export const { setFile } = photoSlice.actions;
+export const { setFile, setPreview } = photoSlice.actions;
 
 export const photoReducer = photoSlice.reducer;
 
 export const selectB64Photo = (state: RootState) => state.photo.B64_photo;
+export const selectPhotoPreview = (state: RootState) => state.photo.preview;